rename package from src to respyratorgui
