+++ Pipfile
@@ -14,3 +14,6 @@ kivy = "*"
 
 [requires]
 python_version = "3.7"
+
+[scripts]
+app = "python main.py"
