+++ respyratorgui/kivygui/app.kv
@@ -15,9 +15,9 @@
 # along with Respyrator.  If not, see <http://www.gnu.org/licenses/>.
 
 #:kivy 1.11.1
-#:import colors src.kivygui.kivyconfig.colors
-#:import imgs src.kivygui.kivyconfig.imgs
-#:import txts src.kivygui.kivyconfig.txts
+#:import colors respyratorgui.kivygui.kivyconfig.colors
+#:import imgs respyratorgui.kivygui.kivyconfig.imgs
+#:import txts respyratorgui.kivygui.kivyconfig.txts
 
 <GuiManager>:
     id: gui_manager
