+++ respyratorgui/kivygui/modesscreen.py
@@ -18,13 +18,13 @@
 # Installed -------------------------------------------------------------------
 from kivy.uix.screenmanager import Screen
 # Coded -----------------------------------------------------------------------
-from src.settings import logapp
+from respyratorgui.settings import logapp
 from .kivyconfig import load_kv
 # Program ---------------------------------------------------------------------
-LOG = 'ParamsScreen:'
+LOG = 'ModesScreen:'
 
 load_kv(__file__)
 
 
-class ParamsScreen(Screen):
+class ModesScreen(Screen):
     pass
