+++ respyratorgui/kivygui/monitoringscreen.py
@@ -18,13 +18,13 @@
 # Installed -------------------------------------------------------------------
 from kivy.uix.screenmanager import Screen
 # Coded -----------------------------------------------------------------------
-from src.settings import logapp
+from respyratorgui.settings import logapp
 from .kivyconfig import load_kv
 # Program ---------------------------------------------------------------------
-LOG = 'ModesScreen:'
+LOG = 'MonitoringScreen:'
 
 load_kv(__file__)
 
 
-class ModesScreen(Screen):
+class MonitoringScreen(Screen):
     pass
